+++ doc/release_notes.rst
@@ -1,6 +1,8 @@
 Release notes
 =============
 
+* Grid optimization for the spline definition implemented (issue #15).
+
 
 *ximpol (0.4.0) - Sat, 09 Jan 2016 10:17:52 -0800*